Found Snapshot crashing bug, hooray!!  (Needed a SectionLock in ObjectStateCache)
Big SectionLock update! See SectionLock.h for some added comments and help.

Ported SectionLock to OSX (should have done before)
Because of this, need to use SWS_Mutex for locks which is an OS independent mutex.
Also, changed the name of SectionLock to SWS_SectionLock to clarify that it's in the SWS project, and goes with SWS_Mutex, etc.
Moved everything into SectionLock.h and deleted SectionLock.cpp.