v2.3.0.3 Issue 498: OSX - Fixed docked SWS window close issues
SWELL IsWindow fails on docked windows.  Justin recommends not using IsWindow at all, but there are a lot of cases where we rely on it for error checking.  Instead of removing all calls and doing internal window state handling I made a SWS_IsWindow function that will check for docked windows on OSX.  It's a bit of a hack, but less risky IMO than rewriting tons of window handling code.