Continuous actions can now be registered in MIDI editor too.
However, there's one catch on OSX - while the action is running, MIDI editor will loose focus.
This is because we can't use keyboard accelerator registered with REAPER to catch keystrokes in MIDI editor (so we focus arrange instead and eat all keystrokes until the action stops) and due to SWELL and OSX in general, we can't use standard win32 tricks to overcome this (hook keyboard procedure or use GetAsyncKeyState/GetKeyboardState) - I do hope Cockos does something about this one day but at least we got it working somehow for now.