Removed exported win32 ReaScript API and actions to simplify title bar
After discussion with Jeffos it was deemed that these API calls could be possibly
dangerous in wrong hands (not to mention some of them weren't implemented properly) :)

So for now, I'll keep these mods private to myself and try to fix them up so they
work properly in all situations. If we decide we want to release this in the future,
please ping me.