remove useless code vulnerable to buffer overflows
A buffer overflow can be triggered by setting an external tool command with a
filename larger than 100 characters. The result of the vulnerable code
(`Tool{1,2}MenuText`) is actually unused since 2011 (commit 447f82db02108402fc039f7dfd411cefa87f3332). This PR removes it entirely.

The other uses of `ExtractFileNameEx` are safe-ish (but fragile) because the
destination buffer is larger than a filename can *currently* be on common
filesystems.