fix empty columns taking no horizontal space on Linux (#1147)
https://i.imgur.com/uQgmQqf.png

This was happening because SWELL-generic skips list items with no text without updating the coordinates of the next item.

SWS was not setting the LVIF_TEXT mask when populating items beyond the first column thinking it was updating an existing item whose text didn't change.

This patch makes SWS always set LVIF_TEXT when adding items to the list even if the string is empty.