fix track spacer handling in wol track height adjustment actions
- SWS/wol: Adjust selected envelope or last touched track height
- SWS/wol: Adjust envelope or track height under mouse cursor

https://github.com/reaper-oss/sws/pull/1787#issuecomment-1934495777

Add GetTrackHeightWithSpacer()

GetTrackHeight() is now balanced to SetTrackHeight(), as previously -- it will get the 'content' portion height and offset of the track in the TCP. If you need the true track offset (- spacerSize) or height (+ spacerSize), use GetTrackHeightAndOffsetWithSpacers().

Updated a couple of Zoom.cpp functions to take the spacer size into account, missed on the first pass.