fix running non-default verbs via CF_ShellExecute
NULL = "The default verb is used, if available. If not, the `open` verb
is used. If neither verb is available, the system uses the first verb
listed in the registry."

Most notably, REAPER associates its file extensions using a "open64" verb which
lead to RPP failing to open using `CF_ShellExecute`.